pyproject.toml
[project]
name = "slms-experiment"
version = "0.1.0"
description = "Experiment for the use of SLMs in Agentic AI."
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "langchain>=1.0.5",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=2.0.5",
    "langgraph>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
]
